requirements.txt
aiofiles==0.8.0
alembic==1.8.1
anyio==3.6.1
atomicwrites==1.4.1
attrs==21.4.0
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.79.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
humps==0.2.2
idna==3.3
iniconfig==1.1.1
Jinja2==3.1.2
keyring==23.7.0
Mako==1.2.1
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
pip==22.2.2
pluggy==1.0.0
psycopg2==2.9.3
psycopg==3.0.16
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.28.1
rsa==4.9
setuptools==62.6.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
tomli==2.0.1
typing-extensions==4.3.0
tzdata==2022.1
urllib3==1.26.11
uvicorn==0.18.1
watchfiles==0.16.1
websockets==10.3
wheel==0.37.1
